Saving Pages
The page title (provided in the uLR) and the form's only field, Body,
are stored in a new Page. The save() method is then called to write the
data to a file, and the client is redirected to the /view/ page.

The value returned by FormValue is of a type string. We must convert
that value to []byte before it will fit into the Page struct. We use
[]byte(body) to perform the conversion